Possible fix for inconsistent import behavior
With a shovel directory and a "shared.py" file imported by other tasks,
the import will fail depending on the order the files are loaded. This
attempts to fix that by making the sort order deterministic, so you can
do your fixup work in __init__.py, or a similar-named file.